only display x axis tick mark for years present
some strange styling was added to odd elements in the x-axis tickmarks
for the homepage stacked area chart. Providing a fixed tick count -- the
number of years of data present displays the correct single tick marks
with no interpolated points between years. I needed to remove the
styling on the odd nodes in order for this to work. I'm not sure why
that styling was present.

closes #3